CocoaPods: Set the correct pURL for an empty package
When the pod command can't find any .podspec file for a pod, an empty package
will be returned for which only the identifier is set. The empty pURL property
in the `Package` object can cause failures later in the Advisor, when requesting
vulnerabilities for packages using the pURL. Therefore, it's necessary to set
the correct pURL for an empty package explicitly, because when using
`Package.EMPTY` it's always empty.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>