SpdxDocumentFile: Prefer VCS information from SPDX over working tree
Prefer the VCS information provided by the SPDX metadata of a package
over the VCS information gathered from the local working tree. Otherwise
it is not possible to define in SPDX that the source code of a package
is located in a different VCS because the local working tree will always
be preferred.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>