Npm: Improve loading of remote package details
Run the external process to lookup remote package details in a
separate coroutine asynchronously. In the cache, only store the
Deferred with the result. This should reduce contention on the cache
map, since it is no longer necessary to wait for the process to
finish.

Extract the caching logic to a new function that invokes the original
getRemotePackageDetails() function. That way subclasses benefit from
the caching mechanism as well.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>