ResolvedLicense: Fix `toResolvedCopyrights()`
When searching the copyright findings for a statement the code was
erroneously using the processed statement instead of the unprocessed
statement. If the processed statement was different to the original
statement this lead to an exception, because the map of original
statements to their findings did not contain an entry with that key.
Also, this error limited the list of findings in the `ResolvedCopyright`
to those which exactly match the processed statement.

This is a fixup for 3e9d11d.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>