clients: Avoid the need for callers to repeat default values
For callers that get client configuration from external sources like a
file it is convenient to just pass along what has been configured, and
if a property was not configured, use `null` to signal that. Allow that
instead of having to repeat default values by changing clients to accept
`null` values and apply the defaults internally themselves.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>