PathExcludeGenerator: Fix path generation on Windows
Running the helper command on Windows should generate the same path
excludes as on non-Windows, so that it does not matter under which
operating system ORT will be run when making use of the generated path
excludes.

However, as Windows' file system is case-insensitive, `File` objects
whose paths only differ in case are considered equal, but for the sake of
generating path exclude patters they must be considered unequal.

That means `File` object must not be stored in `Set`s or be used as keys
in `Map`s as otherwise unwanted deduplication happens. Resolve that by
(mostly) operating on path `String`s instead of `File`s. This also
allows to deduplicate some code by leveraging ORT's existing
`getAllAncestorDirectories()` instead of the `getAncestorFiles()`
extension function.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>