Analyzer: Fix `mustRunAfter` to eventually hang when configured globally
Enabled package managers for which no definition files were found must
not be considered in the dependencies that have to complete before another
package manager is run, because these dependencies never complete as they
are never run. This logic was already implemented for dynamic package
manager dependencies in `determinePackageManagerDependencies()`, but not
for statically configured package manager dependencies. Simply move the
code that makes the static configuration override the dynamic one to
inside `determinePackageManagerDependencies()` so the same logic about
not adding package managers without files to the dependencies kicks in.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>