Rule: Extend hasLabel() to cover labelContains()
Checking a label value without splitting it at commas seems to be an
edge case, because it is unlikedly that users use label values which do
not represent a comma-seprated list but do use commas.

Extend hasLabel() to cover that edge case so that labelContains()
becomes redundant and can be eliminated.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>