NuGetSupport: Separate fetching data from mapping it
Separate responsibilities to simplify working with the code.

The propagation of issues is no more based on exceptions but uses the
return value. With changing getAllPackageData() to fetch the data for all
packages instead of just for a single package, further optimization
potential gets unlocked. It has become easier to parallelize the
(recursive) processing further.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>