NuGetSupport: Eliminate `definitionFile` from constructor params
Previously, the client code of `NuGetSupport` needed to pass the same
`definitionFile` twice, once to the constructor and once to
resolveNuGetDependencies(). Eliminate that by dropping the constructor
parameter.

This make the constructor more lightweight as it no more downloads
things during construction. Further on, the same instance of
NuGetSupport is now used for processing all definition files. As a
result the OkHttpClient instance is used across definition files, which
potentially increases cache hits.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>