init.gradle: Ignore BOM dependencies when comparing dependency trees
When comparing dependency trees, the count of sub dependencies should
not include BOM dependencies, since they are skipped by the
`fetchDependency` function.
This solves a bug where an Android project shipping
'androidx.appcompat:appcompat:1.5.1' would hang indefinitely. This is
because, for instance, 'kotlinx-coroutines-android:1.6.1' depends on
'kotlinx-coroutines-core:1.6.1, kotlin-stdlib-jdk8:1.6.0 and
kotlinx-coroutines-bom:1.6.1'. Since no ORT dependency is created for
'kotlinx-coroutines-bom' because it's a BOM file, it should also not be
taken in account when comparing dependency trees. Otherwise, comparison
would always fail, leading to a cache miss and overgrowth of the latter.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>