Upgrade Detekt to version 1.22
The `ComplexMethod` rule was renamed to `CyclomaticComplexMethod`, and
revisiting the `threshold` revealed that it can be lowered compared to
before (but still needs to be higher than the default of 15).

New rules demand annotated declarations to be on separate lines and
to be separated by blank lines, like

    @SerialName("npm")
    NPM,

    @SerialName("crate")
    CRATE,

While this notation makes sense if the declarations are documented (and
thus already separated by blank lines), for undocumented declarations
(esp. as part of `@Serializable` classes) a more compact notation makes
sense:

    @SerialName("package_name") val packageName: String,
    @SerialName("installed_version") val installedVersion: String,

Disable rules that prevent this notation, as it is already used e.g. in the
`PythonInspector` class. Also, in general disable rules that would require
too many code changes at this point.

Otherwise address the code inspection hints.

For the release notes see [1].

[1]: https://github.com/detekt/detekt/releases/tag/v1.22.0

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>