ResolvedLicense: fix toResolvedCopyrights to identify the same copyright in multiple files.
Fix a regression when a copyright is present in multiple files in the same
repository.
The expected behavior is that the all findings of the copyright are returned,
if they have different paths.
The previous behavior was that only the last finding was kept, as
'statementToFinding' is built using 'associateBy' which "If any two elements
would have the same key returned by keySelector the last one gets added to the
map."

Signed-off-by: Stefano Bennati <stefano.bennati@here.com>