ScannerWrapper: Inline `ScannerDetails` to deduplicate the `name`
Having `ScannerDetails` as part of `ScannerWrapper` creates a
recursion pitfall for bootstrappable `CommandLinePathScannerWrapper`s as
they must not access the `ScannerDetails` to get the `name` during
bootstrapping as that would also trigger getting the `version` which is
not known before bootstrapping succeeded.

To avoid running into that pitfall, inline the `version` and
`configuration` properties, and only create `ScannerDetails` when
needed.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>