ConfigCommand: Add a simple conversion from HOCON to YAML
It is convenient to be able to use a current version of ORT to assist
users with migrating from HOCON to YAML configuration despite HOCON
support being removed as of 8db2c23. Thus, add an option that performs a
simple conversion on a pure search and replace basis, so it can handle
template values that would otherwise result in deserialization issues.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>