Remove the feature to bootstrap scanners
Align with the semantics of analyzer tools, which is that any external
tools need to be present on the host running ORT. This allows to remove
quite some complexity from the scanner code. Bootstrapping was barely
used anyway, with ScanCode de facto being the only scanner used, and
ScanCode also being installed in the `Dockerfile`.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>