GoMod: Fix support for replace directives
Change `getModuleInfos` to return a map which associates the module
paths to the related module info. The keys of the map contain both, the
paths of replaced modules and those of their replacements. This fixes
support for replace directives which do not only replace the version but
also the path of a module. Previously, this caused a
`NoSuchElementException` because module infos were only associated to
the path of the replacing module, but the output of `go mod graph` also
contains the paths of replaced modules.

Fixes #5604.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>