Gradle: Ignore version constraint dependencies
The `DependencyResult` class has a `constraint` property which indicates
that it does not represent a dependency of its parent but defines a
version constraint which was used to resolve the version of the parent.
Therefore, do not add these nodes to the dependency tree.

Especially for Android projects this can dramatically reduce the size of
the dependency tree, as demonstrated by the reduced size of the expected
result for the "Cyclic dependencies over multiple libraries can be
handled" test and the Pub Flutter test.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>