model: Fix the `Set` contract for data classes that implement `Comparable`
Kotlin's data classes have auto-generated `equals()` functions that take
all primary properties into account. When implementing the `Comparable`
interface for those classes, i.e. to give them a defined order in a
`SortedSet`, special attention needs to be put on `compareTo()` to
return 0 if and only if `equals()` returns true. This is required because
"a sorted set performs all element comparisons using its compareTo
method, so two elements that are deemed equal by this method are, from the
standpoint of the sorted set, equal" [1]. This means that sorted sets are
considered equal if for all their respective elements `compareTo()`
returns 0. So if `compareTo()` is implemented to not consider all
properties for an equality check, this can lead to unexpected results, for
example when using Kotest's data class matchers.

Fix this and obey the general contract of the `Set` interface by taking
all properties into account if the properties to sort on are equal. As
the ordering does not matter if not all properties are equal but the
properties to sort on are, arbitrarily return 1 from `compareTo()` in
that case.

[1]: https://docs.oracle.com/javase/8/docs/api/java/util/SortedSet.html

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>