analyzer: Stop using a `SortedSet` in `ProjectAnalyzerResult`
The `ProjectAnalyzerResult` is only serialized as part of analyzer
functional tests when comparing to expected results. So in memory, there
is no need for the `packages` to be a `SortedSet`. Change the code to
only sort `packages` on serialization via a Jackson converter to have a
defined order when comparing to expected results.

This lays the fundation for getting rid of `Comparable` implementations
in various classes. These implementations were only added for sorted
output during serialization, but they break the `Set` contract as
`compareTo() == 0` is not aligned with `equals() == true`, also see [1].

[1]: https://docs.oracle.com/javase/8/docs/api/java/util/SortedSet.html

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>