Dockerfile: Ensure that the `.ort` data directory exists
It is common practice to mount a host-side ORT config directory into the
Docker container at `/home/ort/.ort/config`. While Docker infers the
permissions for the `config` directory from the corresponding directory
on the host, for the intermediate `.ort` directory (which did not exist
in the container prior to this change) there is no corresponding
directory on the host, so Docker uses `root:root` ownership. Fix the
ownership to be `ort:ort` by ensuring that the `.ort` directory exists
in the container.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>