Docker: Use Java 17 by default to build and run ORT
This basically reapplies the commit reverted in 960f702 as it was agreed
in the community meeting that by now Java 17 should be the default [1],
but the Java version should be configurable via a Docker build argument.
Users that need to analyze projects incompatible with Java 17 should use
`--build-arg JAVA_VERSION=11` with `docker build` to restore the
previous behavior.

[1]: https://github.com/oss-review-toolkit/ort/wiki/Community-Meeting#2022-12-15-last-meeting-before-holiday-season

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>