HttpFileStorage: Increase the connect timeout from 10 to 30 seconds
Executing `ort scan` inside Docker on a GitHub Runner with the storages
for scan results, provenances and file archives configured as HTTP file
storages, all pointing to a certain JFrog instance runs into [1]
consistently. Many times accessing the storage succeeds but sometimes it
fails. Given a medium sized analyzer result as scanner input, the
problem always reproduces.

While the root cause is not understood, increasing the connect timeout
of the HTTP client fixes the issue. Use that as a work around without
changing the default HTTP clients configuration to keep the impact of
this change limited.

Fixes: #6261.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>