Gradle: Replace ':' in "absolute" project names
As ORT uses ':' as a separator in project / package identifiers, the
names themselves must not contain ':'. However, this may happen for
yet unclear reasons, and projects may use an "absolute path" as the
project name where Gradle uses ':' to separate projects in a subproject
hierarchy. Avoid exceptions in those cases by following the convention
to use the path to the definition file as a fallback for the project name.

[1]: https://docs.gradle.org/current/userguide/intro_multi_project_builds.html#sec:executing_tasks_by_fully_qualified_name

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>