feat(Conan): Add support for Conan lockfiles
Require a lockfile if a path to the lockfile is configured in
`config.yml`.

Only lockfiles located in the same directory as the definition file are
supported.

The test for Conan lockfiles is configured to only run on Linux as
`package id` in the lockfile differs depending on operating system.

Signed-off-by: Oleksandr Yarosevych <fixed-term.oleksandr.yarosevych@bosch.io>