AnalyzerCommand: Stop using the `FallbackPackageCurationProvider`
Stop using the `FallbackPackageCurationProvider` because it can lead to
unexpected behavior if a new curation is added and as a result a
curation from a fallback provider is not used anymore.

For example, if a fallback provider provides a curation that fixes the
VCS URL of a package, this curation would not be used anymore if a new
curation for the same package is added to a provider with higher
priority, even if it curates a different property than the VCS URL. This
means that the new curation would have to contain the VCS URL which
might only become apparent after running the analyzer and scanner again.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>