AnalyzerCommand: Create package curation providers from config file
Use the `PackageCurationProviderFactory` to create the package curation
providers that are configured in the config file and remove the now
unused command line options to enable package curation providers.

The only exception is currently the provider that reads package
curations from the repository configuration, this will be handled in a
later commit.

Remove the test from `OrtMainFunTest` that tests the removed
`--package-curations-file` option and instead apply curations in the
"Analyzer creates correct output" test.

Fixes #6055.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>