CreateAnalyzerResultCommand: Migrate to provenance based scan storage
The command operates on a package based scan result storage which has
become legacy. So, operate on a provenance based scan result storage
instead which has become the default way of storing scan results.

The most recent resolved provenance now is no more determined via the
scan start time, but via the row id of the provenance resultion result.
This is much faster, because it works without accessing the scan result
JSONB column.

Note: The primary use case this command targets is re-scanning packages
prior to upgrading ScanCode so that the "cache" is warmed-up.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>