JitPack: Switch back to OpenJDK 11
Switch back to OpenJDK 11 for building the JitPack artifacts. Otherwise
it is not possible to use those artifacts to build plugins that still
need to run on Java 11, for example when using a customer Docker image
with Java 11 instead of 17.

Trying to build with Java 17 and target 11 results in this error
message:

Cannot inline bytecode built with JVM target 17 into bytecode that is
being built with JVM target 11. Please specify proper '-jvm-target'
option.

This partly reverts 171fa3d.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>