build: Use GraalVM's native build tools Gradle plugin
The advantage of GraalVM's native build tools plugin [1] over Palantir's
plugin [2] is that reflection metadata is automatically downloaded and
Gradle's native toolchain mechanism [3] is used (via the Foojay Toolchains
Plugin [4]) to bootstrap GraalVM. By default, it uses the latest GraalVM
release for the language version of the JVM that runs the Gradle build.

A small drawback of this approach is that Gradle does not maintain
symbolic links [5] when unpacking the GraalVM distribution, so these
need to be restored manually.

Start fresh with empty GraalVM native image build arguments.

[1]: https://graalvm.github.io/native-build-tools/latest/gradle-plugin.html
[2]: https://github.com/palantir/gradle-graal
[3]: https://docs.gradle.org/current/userguide/toolchains.html
[4]: https://github.com/gradle/foojay-toolchains
[5]: https://github.com/gradle/gradle/issues/3982

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>