refactor(Advisor): Split the `advise` function
The advice providers only need a set of packages as input, therefore
split out the logic that queries the advice providers into a separate
`advise` function that only takes a set of packages as input.

The new function makes it easier to use the `Advisor` programmatically
because it does not require to build a complete `OrtResult`.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>