fix(commitlint): Parse breaking change headers correctly
Before this change, using a header like "fix!: Something" caused the
following commitlint errors:

✖   subject may not be empty [subject-empty]
✖   type may not be empty [type-empty]

It seems that the default header pattern does not recognize the
exclamation mark. To fix this, use the parser options from the
"conventional-changelog-conventionalcommits" template [1]. Using the
package as a parser preset would require that the NPM package is
installed, therfore copy the values from [2] instead.

[1]: https://github.com/conventional-changelog/conventional-changelog/tree/master/packages/conventional-changelog-conventionalcommits
[2]: https://github.com/conventional-changelog/conventional-changelog/blob/bfe3bf1c49d4a125474b398b2d304749fd3b56c7/packages/conventional-changelog-conventionalcommits/parser-opts.js

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>