fix(VersionControlSystem): Correctly check for availability
The test whether a VersionControlSystem implementation based on a
command line tool is available was only working for classes
implementing the CommandLineTool interface. If the implementation used
a different mechanism to invoke external commands, it was considered
to be always available. This could cause IOExceptions to be thrown
when resolving VCS URLs due to the invocation of commands not
available on the path.

Fix this by adding a constructor parameter accepting a CommandLineTool
to VersionControlSystem. Subclasses that manage their command line
tools differently can pass here the correct object. Use this parameter
in the implementations for CVS and Mercurial, which have been affected
by this issue.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>