feat(Gradle): Add support for composite builds
Composite builds are a feature of Gradle [1].
This commit addresses an issue when a build is dependent on a subproject
of an included build: ORT cannot generate valid package identifiers for
such subprojects because it has only the included build name and the
subproject name, lacking the group ID and version of the included build.
Indeed, the Gradle Tooling API does not seem to provide information
about included builds when using `getModel()` as ORT does (see also [2]
and [3]).
Therefore, this commit adds support for composite build by using Gradle
internal API. A functional test is also provided.

[1]: https://docs.gradle.org/current/userguide/composite_builds.html
[2]: https://github.com/eclipse/buildship/issues/965#issuecomment-574653434
[3]: https://stackoverflow.com/a/70688166/5321324

Co-authored-by: Marcel Bochtler <marcel.bochtler@bosch.io>
Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>