test(assets): Fix-up `ort-analyzer-result.yml`
The VCS info `OrtResult.repository.vcsProcessed` must always be
normalized. In case it is not, ORT runs into an exception [1] when
calling `OrtResult.getDefinitionFilePathRelativeToAnalyzerRoot()`.

This in fact happens for this test asset. Change the VCS URL to the
corresponding normalized one to fix the problem.

Also update the URLs in VCS processed of each respective project for
consistency, even though not necessary to fix the mentioned exception.

[1] https://github.com/oss-review-toolkit/ort/blob/a0d9181b2c7d2cf537611ec476fb016bb3a95704/model/src/main/kotlin/OrtResult.kt#L246

Signed-off-by: Frank Viernau <frank_viernau@epam.com>