test(assets): Fix-up `gradle-all-dependencies-expected-result.yml`
The asset contains placeholder strings of the form `<REPLACE_...>`.
These placeholder are unused, because they never get actually replaced.
Replace the placeholder with real world values, choosing an arbitrary
ort revision as revision.

Previously, using this asset lead to crashes as soon as the VCS info
gets parsed into a URL which is now fixed.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>