fix(CuratedPackage)!: Eliminate toUncuratedPackage()
The implementation of `toUncuratedPackage()` does not work for `null`
values in all cases due to lack of data, see also #5637.

The representation of `OrtResult` has been changed recently to store the
original / uncurated packages instead of the curated ones. So, make use
of that new representation to always obtain the correct uncurated
package.

Fixes #5637.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>