fix(PackageCuration): Remove semver coercing
Neither the package's nor the curation's version are guaranteed to use
valid semvers. Semver4j's `coerce()` function creates a valid semver in
cases when this is not wanted, e.g. 1.2.3.4 is parsed as 1.2.3 and
therefore a package curation for 1.2.3.9 would be seen as applicable.

Remove the coercing to valid semvers to avoid applying incorrect package
curations.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.com>