fix(Gradle): Fix regression when project uses Gradle < 7.2.0
In a previous commit ([1]), support for Composite build was added using
Gradle Internal API. Unfortunately, this introduced a regression for
all projects using Gradle < 7.2.0 which could not be analyzed anymore.
This commit fixes the regression by using reflection to call the
internal API.
For projects running Gradle < 7.2.0 and using the composite build
feature, the behavior is restored to the previous one i.e. it will
throw a `NullPointerException` when accessing the group ID of a
subproject from an included build.

Solves https://github.com/oss-review-toolkit/ort/issues/6419.

[1]: https://github.com/oss-review-toolkit/ort/commit/6cce314c695e9e4071ef3f8a7956b2683cf2b9d5

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>