feat(OrtResult)!: Associate package curations with provider IDs
Extend `ResolvedConfiguration` to associate the curations with the
ID of the package curation provider to enable tracability of curations
back to the provider. The data model is adjusted as agreed in [1],
while adding `config` and `type` of the provider is left for a future
change.

This implementation also is a first step towards use cases involving:

1. Replacing the curations for a given provider ID with the given ones.
2. Re-resolve curations only for a particular provider ID.

Both are left as TODO for future changes to limit the size of this
change, while for 1. a TODO comment is left in the code.

Fixes #5668.

[1]: https://github.com/oss-review-toolkit/ort/pull/6456#discussion_r1107282851

Signed-off-by: Frank Viernau <frank_viernau@epam.com>