refactor(Analyzer)!: Improve package curation ordering
Consistently order the package curations or the providers respectively
by highest-priority first. This avoids confusion compared to dealing
with inconsistent ordering. This way only the logic which actually
applies the curations needs to reverse the curations.

Add or update the KDoc of all releated functions and properties so that
the order is always specified.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>