refactor(PackageCurationProvider)!: Make `getCurationsFor()` return a list
Previously, `getCurationsFor()` returned a map keyed by the `Identifier` of
the `Package` that was used to request the curation. For curations that
match multiple ids, e.g. because of version ranges, this resulted in the
same curation being returned multiple times for different ids.

To avoid these duplicates and because all callers of `getCurationsFor()`
anyway call `flatten()` on the map's values only, drop the keys by simply
returning a list instead.

Note that the `PackageCuration.isApplicable()` check needs to happen
independently of that before writing curations to the ORT result as there
is no guarantee that all curation provider implementations only return
applicable curations in the ORT sense.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>