refactor(VcsType): Avoid the creation of new instances for known types
Previously, new `VcsType` instances were created with the same list of
aliases as for known types, resulting in a waste of memory. Change the
code to replace the single-type constructor with a factory function that
returns references to known types instead, and creates new `VcsType`
instances only for unknown types / a non-matching alias.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>