feat(OrtResult): Handle curations when replacing the repository config
Several CLI commands use this function to override the repository
configuration of an `OrtResult` and afterwards operate on that
"patched" version of it.

Back when package curations were added to the repository configuration
(behind a feature flag) replaceConfig() hasn't been updated to
gracefully handle potential changes in curations.

Add that missing handling so that the effect of a locally changed
`ort.yml` can be seen when used with the respective CLI commands.

Note: The original `OrtResult` contains the provider with the identifier
REPOSITORY_CONFIGURATION_PROVIDER_ID if and only if package curations in
repository configurations had been enabled during the creation of the
ORT file, see also `OrtConfiguration.enableRepositoryPackageCurations`.
Deliberately only set the curations if that feature was enabled during
creation of the original ORT file in order to not start dealing with
OrtConfiguration changes in between the particular stages which only
complicates things.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>