fix(DirectoryStash): Make the temporary directory start with a dot
The temporary directory is a sibling of the original directory, which
means they share the same parent directory. If, like in the case of
the `GoMod` package manager, the common parent directory is searched
[1], this means that stashed files undesirably may still be found.

Address that specificially for the case of `GoMod` by marking the
directory as "hidden" by starting its name with a dot, as Go ignores
such directories [2].

Fixes #6519.

[1]: https://github.com/oss-review-toolkit/ort/blob/fec5dc5/analyzer/src/main/kotlin/managers/GoMod.kt#L102-L103
[2]: https://github.com/golang/go/issues/30058#issuecomment-459888562

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>