refactor(GoMod): Disable any GOPROXY
This ensures that the `.info` files for all modules, which
get created by the Go command line tools under '$GOPATH/pkg/mod',
contain the VCS info. When using a Go proxy this is not guaranteed and
depends on the actual proxy instance, see [1]. This comes at the price
of increased execution time.

As the user's environment may likely be configured to use a GOPROXY,
which is the default, use an empty directory to not mix-up newly cached
files with already existing ones which might have been created using a
Go proxy and therefore may lack the VCS info.

This prepares for using the Go tooling to determine the VCS info instead
of using ORT's partial reimplementation of that.

[1]: https://github.com/golang/go/issues/44742#issuecomment-1210986778
[2]: https://github.com/golang/go/issues/18387

Signed-off-by: Frank Viernau <frank_viernau@epam.com>