deps(reporter-web-app): Backward-incompatible Yarn updates
Backward-incompatible updates might need some code changes but in this
commit the dependencies that are updated don't show any errors and the
app runs successfully after the updates.

The updates are needed to enhance the security and to have the
application up-to-date.

Changed the test range in WebAppReporterFunTest.kt from
2070000L..2090000L to 2140000L..2160000L as the file size gets bigger
after the updates.

Signed-off-by: Menatalla Elnemr <menatalla.elnemr@hella.com>