feat(MavenSupport): Improve the caching logic for empty remote artifacts
Previously, failures to look up remote artifacts were unconditionally
cached as empty artifacts. While it is hard to reliably determine the
root cause of a failed lookup, in the case that all candidate URLs do
not end with a known extension it is very likely that ORT ran into a bug
in the context of [1]. Ease debugging such cases by not anymore
caching failed lookups due to missing file extensions, and generally
improve the logging when returning an empty remote artifact.

[1]: https://github.com/oss-review-toolkit/ort/issues/4694

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>