feat(detekt): Allow deep nesting in OrtPackageNaming rule
Extend the `OrtPackageNaming` rule to allow an arbitrary depth of nested
modules. Before, only one level of nesting for Gradle modules was
supported. This prepares for an upcoming change which introduces one
more level of nesting in the Gradle module structure.

To keep the code simple, already allow an arbitrary depth of nesting
even though it is not required yet.

While at it, ignore dashes in module names also for parent modules.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>