feat(NoticeTemplateReporter)!: Change determining the output filename
The `NoticeTemplateReporter` used "NOTICE_${templateName}" as the
output filename. Remove the notice file specific "NOTICE_" prefix to
prepare for generalizing the `NoticeTemplateReporter` to a
`PlainTextTemplateReporter`. So, given a template named
'my-summary.md.ftl' the reporter now produces 'my-summary.md'.

The filenames of the built-in notice templates are now prefixed with
"NOTICE_" to keep that prefix for the output files. While at it, the
suffix is changed from lowercase to uppercase for consistency. Instead
of `NOTICE_default` the reporter now produces `NOTICE_DEFAULT`.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>