deps(reporter-web-app): Update React packages
React related pages that are updated are "react","react-app-polyfill",
"react-dom","react-redux","react-scripts","react-syntax-highlighter" and
"@testing-library/react".

The only packages that needed changes in the code are "react",
"react-dom" and "react-scripts".

In the "react-dom" V18.2.0 the rendering of the app differs than
"react-dom" V17.0.2 so This what required to change the code in index.js
.

Also "react-dom" requires "react" package to be up to date.

Fix the commenting style of the first line in index.css as
"react-scripts" had a problem with it because of CSS minification.

Signed-off-by: Menatalla Elnemr <menatalla.elnemr@hella.com>